+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - overwriteRequestBody - added support for type "x-www-form-urlencoded" (#325)
 - AssignVariables - Added support for null values for response body (#414)
 - ContentTests - Added support for null values for response body (#414)
+- ContentTests - Added support to lookup a key within an array of objects by using a * wildcard (#417)
 
 ## v1.18.1 - (2022-09-30)